Manually Adding Researcher Folder
Due to issues coming up with me being unable to work directly in the project due to claims of a missing module I had to pull together my assets and build the supporting character in an old copy of the project. 
If no conflicts come up, there should be a folder in Characters called HeadResearcher, and inside that is a single researcher blueprint and its accompanying assets. 
There is also the start to a blueprint interface and widget in the event we want to try and go through with making this character a hint-giver.